+++ package.json
@@ -7,7 +7,7 @@
     "test": "tests"
   },
   "scripts": {
-    "test": "NODE_ENV=test jest",
+    "test": "NODE_ENV=test jest --detectOpenHandles",
     "start": "nodemon server.js",
     "start:prod": "NODE_ENV=production nodemon server.js"
   },
