+++ server.js
@@ -2,6 +2,9 @@ const dotenv = require("dotenv");
 const { default: mongoose } = require("mongoose");
 const app = require("./app");
 
+const { MongoMemoryServer } = require("mongodb-memory-server");
+let mongod = null;
+
 //manage uncaught exception
 process.on("uncaughtException", (err) => {
   console.log("Uncaught Exception. Shutting down...");
@@ -11,33 +14,59 @@ process.on("uncaughtException", (err) => {
 
 dotenv.config({ path: "./config/config.env" });
 
-//replace password with
-// const DB = process.env.DATABASE.replace(
-//   "<PASSWORD>",
-//   process.env.DATABASE_PASSWORD
-// );
+const connectDB = async () => {
+  try {
+    //replace password with
+    let DB = process.env.DATABASE.replace(
+      "<PASSWORD>",
+      process.env.DATABASE_PASSWORD
+    );
+    if (process.env.NODE_ENV === "test") {
+      mongod = await MongoMemoryServer.create();
+      DB = mongod.getUri();
+    }
 
-// mongoose
-//   .connect(DB, {
-//     useNewUrlParser: true,
-//   })
-//   .then(() => {
-//     console.log("DB connection successful!");
-//   });
-//.catch(err => console.log("ERROR")); if password is wrong print error
+    const conn = await mongoose
+      .connect(DB, {
+        useNewUrlParser: true,
+      })
+      .then(() => {
+        console.log("DB connection successful!");
+      });
+  } catch {
+    (err) => console.log("ERROR"); //if password is wrong print error
+  }
+};
+
+const disconnectDB = async () => {
+  try {
+    await mongoose.connection.close();
+    if (mongod) {
+      await mongod.stop();
+    }
+  } catch (err) {
+    console.log(err);
+    process.exit(1);
+  }
+};
 
 const port = process.env.PORT || 3000;
 
-const server = app.listen(port, () => {
-  console.log(`App running on port ${port}`);
-});
-
-//manage other errors
-process.on("unhandledRejection", (err) => {
-  console.log("Unhandled Rejection. Shutting down...");
-  console.log(err.name, err.message);
-  //before closing server and after the app
-  server.close(() => {
-    process.exit(1); //0 is for success and 1 is for errors
+function createHttpserver(port) {
+  const app = require("./app");
+  const server = app.listen(port, () => {
+    console.log(`App running on port ${port}`);
   });
-});
+  return server;
+}
+//manage other errors DA RIMETTERE
+// process.on("unhandledRejection", (err) => {
+//   console.log("Unhandled Rejection. Shutting down...");
+//   console.log(err.name, err.message);
+//   //before closing server and after the app
+//   server.close(() => {
+//     process.exit(1); //0 is for success and 1 is for errors
+//   });
+// });
+//if (process.env.NODE_ENV !== "test") createHttpserver(port);
+module.exports = { connectDB, disconnectDB, createHttpserver };
