--- tests/mockDB.js
@@ -1,44 +1,0 @@
-const mongoose = require("mongoose");
-const dotenv = require("dotenv");
-
-dotenv.config({ path: "./config/config.env" });
-
-const { MongoMemoryServer } = require("mongodb-memory-server");
-let mongod = null;
-
-const connectDB = async () => {
-  try {
-    let dbUrl =
-      "mongodb+srv://avidia:IpMqeX374qQfhlcw@cluster0.zxowbqr.mongodb.net/webbees?retryWrites=true&w=majority";
-    if (process.env.NODE_ENV === "test") {
-      mongod = await MongoMemoryServer.create();
-      dbUrl = mongod.getUri();
-    }
-
-    const conn = await mongoose.connect(dbUrl, {
-      useNewUrlParser: true,
-      //useUnifiedTopology: true,
-      //useFindAndModify: false,
-    });
-
-    console.log(
-      `MongoDB connected: ${(conn.connection.host, conn.connection.port)}`
-    );
-  } catch (err) {
-    console.log(err);
-    process.exit(1);
-  }
-};
-const disconnectDB = async () => {
-  try {
-    await mongoose.connection.close();
-    if (mongod) {
-      await mongod.stop();
-    }
-  } catch (err) {
-    console.log(err);
-    process.exit(1);
-  }
-};
-
-module.exports = { connectDB, disconnectDB };
